refactor(client): regroup Analysis Builder hooks by responsibility (#914)
Replace the ~1,050-line useAnalysisBuilder coordinator + 8 feature-sliced
sub-hooks with three responsibility-grouped hooks composed behind the
unchanged public useAnalysisBuilder facade, along a strictly acyclic
State -> Query -> Effects data flow:

- useAnalysisState   store reads/derivation, query-spec building, combined
                     fields, chart config + availability + mode-aware setters,
                     per-mode state + server queries, validation, UI state.
                     No dependency on execution.
- useAnalysisQuery   execution only: the 5 TanStack hooks, mode routing, skip
                     flags, results/loading/error/debug, hasDebounced.
- useAnalysisEffects init/URL parsing, AI generation (direct store access),
                     share, chart-type auto-switch (sole hasDebounced consumer),
                     and external onQueryChange/onChartConfigChange callbacks.

AI + share fold into Effects: AnalysisBuilder/index.tsx no longer calls
useAnalysisAI/useAnalysisShare or carries their glue; it reads AI/share state
and actions off the facade. The facade return shape is unchanged (additive
only) - AI/share fields are now wired to the real Effects outputs.

Pure helpers stay extracted (analysisQueryExecutionModes, buildCubeQuery,
multi-query validators). Deletes the 8 old sub-hook files and the 6 orphaned
sub-hook unit-test files (they import the deleted modules); the #914 safety
net covers the regrouped behaviour at the facade level.

The store and its slices are untouched (hooks-only refactor); store slice
restructuring is tracked separately in #919.

test:client (158 files / 5864 tests), typecheck, and lint all green.
