+++ src/client/components/AIAssistant/index.ts
@@ -2,8 +2,8 @@
  * AI Assistant utilities export
  *
  * Note: AIAssistantModal has been deprecated and removed.
- * The inline AI panel in AnalysisBuilder uses useAnalysisAI hook which
- * depends on these utilities.
+ * The inline AI panel in AnalysisBuilder is driven by useAnalysisEffects
+ * (the AI section of the useAnalysisBuilder facade), which depends on these utilities.
  */
 
 export * from './types.js'
